# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CSV Analyser API",
        "description": "High-performance FastAPI backend for CSV data analysis, Open Icecat product enrichment, preview, and visualizations with MongoDB persistence.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/csv/upload": {
            "post": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Upload Csv",
                "description": "Upload a CSV file, parse it, and return metadata with preliminary statistical summary.\nAlso saves analysis record into MongoDB if connected.",
                "operationId": "upload_csv_api_csv_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_csv_api_csv_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSVSummaryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/datasets": {
            "get": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "List Datasets",
                "description": "List uploaded dataset records stored in MongoDB.",
                "operationId": "list_datasets_api_csv_datasets_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 50,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": true
                                    },
                                    "title": "Response List Datasets Api Csv Datasets Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/metadata": {
            "get": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Get Metadata",
                "description": "Get detailed dataset metadata (row/col count, memory usage, duplicate count, null count).",
                "operationId": "get_metadata_api_csv__file_id__metadata_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSVMetadata"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/summary": {
            "get": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Get Summary",
                "description": "Get full statistical summary, column-level distributions, and correlation matrix.",
                "operationId": "get_summary_api_csv__file_id__summary_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSVSummaryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/preview": {
            "get": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Get Preview",
                "description": "Get paginated rows of the dataset.",
                "operationId": "get_preview_api_csv__file_id__preview_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "Page number",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "Page number"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "description": "Number of rows per page",
                            "default": 20,
                            "title": "Page Size"
                        },
                        "description": "Number of rows per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSVPreviewResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/filter": {
            "post": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Filter Dataset",
                "description": "Filter, sort, and paginate rows based on column conditions.",
                "operationId": "filter_dataset_api_csv_filter_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CSVPreviewResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/chart": {
            "post": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Get Chart Data",
                "description": "Generate aggregated data points for visual charts (bar, line, pie, histogram).",
                "operationId": "get_chart_data_api_csv_chart_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChartDataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChartDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/download": {
            "get": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Download File",
                "description": "Download the CSV file.",
                "operationId": "download_file_api_csv__file_id__download_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}": {
            "delete": {
                "tags": [
                    "CSV Analysis"
                ],
                "summary": "Delete Dataset",
                "description": "Delete dataset record from MongoDB and local storage.",
                "operationId": "delete_dataset_api_csv__file_id__delete",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_csv_api_csv_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_csv_api_csv_upload_post"
            },
...
            "CSVMetadata": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "row_count": {
                        "type": "integer",
                        "title": "Row Count"
                    },
                    "column_count": {
                        "type": "integer",
                        "title": "Column Count"
                    },
                    "file_size_bytes": {
                        "type": "integer",
                        "title": "File Size Bytes"
                    },
                    "columns": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Columns"
                    },
                    "column_types": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Column Types"
                    },
                    "memory_usage_kb": {
                        "type": "number",
                        "title": "Memory Usage Kb"
                    },
                    "total_null_cells": {
                        "type": "integer",
                        "title": "Total Null Cells"
                    },
                    "duplicate_rows": {
                        "type": "integer",
                        "title": "Duplicate Rows"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "filename",
                    "row_count",
                    "column_count",
                    "file_size_bytes",
                    "columns",
                    "column_types",
                    "memory_usage_kb",
                    "total_null_cells",
                    "duplicate_rows"
                ],
                "title": "CSVMetadata"
            },
            "CSVPreviewResponse": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "total_rows": {
                        "type": "integer",
                        "title": "Total Rows"
                    },
                    "columns": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Columns"
                    },
                    "rows": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Rows"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "total_rows",
                    "columns",
                    "rows",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "CSVPreviewResponse"
            },
            "CSVSummaryResponse": {
                "properties": {
                    "metadata": {
                        "$ref": "#/components/schemas/CSVMetadata"
                    },
                    "columns_summary": {
                        "items": {
                            "$ref": "#/components/schemas/ColumnSummary"
                        },
                        "type": "array",
                        "title": "Columns Summary"
                    },
                    "correlations": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "additionalProperties": {
                                        "anyOf": [
                                            {
                                                "type": "number"
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    },
                                    "type": "object"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Correlations"
                    }
                },
                "type": "object",
                "required": [
                    "metadata",
                    "columns_summary"
                ],
                "title": "CSVSummaryResponse"
            },
            "ChartDataRequest": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "x_column": {
                        "type": "string",
                        "title": "X Column"
                    },
                    "y_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Y Column"
                    },
                    "chart_type": {
                        "type": "string",
                        "title": "Chart Type",
                        "description": "bar, line, scatter, pie, histogram, box",
                        "default": "bar"
                    },
                    "aggregation": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aggregation",
                        "description": "count, sum, avg, min, max"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit",
                        "default": 50
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "x_column"
                ],
                "title": "ChartDataRequest"
            },
            "ChartDataResponse": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "chart_type": {
                        "type": "string",
                        "title": "Chart Type"
                    },
                    "x_column": {
                        "type": "string",
                        "title": "X Column"
                    },
                    "y_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Y Column"
                    },
                    "labels": {
                        "items": {},
                        "type": "array",
                        "title": "Labels"
                    },
                    "values": {
                        "items": {},
                        "type": "array",
                        "title": "Values"
                    },
                    "series": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Series"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "chart_type",
                    "x_column",
                    "labels",
                    "values"
                ],
                "title": "ChartDataResponse"
            },
            "ColumnSummary": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "dtype": {
                        "type": "string",
                        "title": "Dtype"
                    },
                    "null_count": {
                        "type": "integer",
                        "title": "Null Count"
                    },
                    "null_percentage": {
                        "type": "number",
                        "title": "Null Percentage"
                    },
                    "unique_count": {
                        "type": "integer",
                        "title": "Unique Count"
                    },
                    "sample_values": {
                        "items": {},
                        "type": "array",
                        "title": "Sample Values",
                        "default": []
                    },
                    "stats": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Stats",
                        "default": {}
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "dtype",
                    "null_count",
                    "null_percentage",
                    "unique_count"
                ],
                "title": "ColumnSummary"
            },
...
            "FilterRequest": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "conditions": {
                        "items": {
                            "$ref": "#/components/schemas/FilterCondition"
                        },
                        "type": "array",
                        "title": "Conditions",
                        "default": []
                    },
                    "sort_by": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sort By"
                    },
                    "sort_ascending": {
                        "type": "boolean",
                        "title": "Sort Ascending",
                        "default": true
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page",
                        "default": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size",
                        "default": 20
                    }
                },
                "type": "object",
                "required": [
                    "file_id"
                ],
                "title": "FilterRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        }
    }
}